+++ Editor/SceneViewToolbar.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor;
 using System;
+using GameplayIngredients.Comments.Editor;
 
 namespace GameplayIngredients.Editor
 {
@@ -108,6 +109,17 @@ namespace GameplayIngredients.Editor
                         EditorWindow.GetWindow<CheckWindow>();
                     }
 
+                    // Comments Window
+                    GUILayout.Space(16);
+                    if (GUILayout.Button(Contents.commentsWindow, EditorStyles.toolbarButton))
+                    {
+                        CommentsWindow.Open();
+                    }
+                    if (GUILayout.Button(Contents.addComment, EditorStyles.toolbarButton))
+                    {
+                        SceneCommentEditor.CreateComment();
+                    }
+
                     // Custom Code here
                     if (OnSceneViewToolbarGUI != null)
                         OnSceneViewToolbarGUI.Invoke(sceneView);
@@ -115,7 +127,6 @@ namespace GameplayIngredients.Editor
                     GUILayout.FlexibleSpace();
                     // Saving Space not to overlap view controls
                     GUILayout.Space(96);
-
                 }
             }
 
@@ -157,6 +168,8 @@ namespace GameplayIngredients.Editor
             public static GUIContent linkGameView;
             public static GUIContent linkGameViewCinemachine;
             public static GUIContent checkWindow;
+            public static GUIContent commentsWindow;
+            public static GUIContent addComment;
 
             static Contents()
             {
@@ -172,6 +185,11 @@ namespace GameplayIngredients.Editor
 
                 checkWindow = new GUIContent(EditorGUIUtility.IconContent("Valid"));
                 checkWindow.text = "Check";
+
+                commentsWindow = new GUIContent(EditorGUIUtility.IconContent("console.infoicon.inactive.sml"));
+                commentsWindow.text = "Comments";
+
+                addComment = new GUIContent("+");
             }
         }