+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added Rigs to **Ingredients Explorer** : that view summarizes currently loaded rigs, and groups them in update groups.
 * Added abstract `PingableEditor` class for MonoBehaviours that can be Pinged to stand out in the inspector.
 * Added Help items in Menu, Added HelpURL class attributes
+* Feature: Scene Comments
 
 ## 2020.2.2